Revert "Improve keypresses within list blocks"
This reverts commit e9d487bd69ff2962068ccdcb51e77a2872bab62f.
