+++ src/blocks/scribe-plugins/scribe-text-block-plugin.js
@@ -1,11 +1,6 @@
 "use strict";
 
-var {
-  getTotalLength,
-  isAtStart,
-  isAtEnd,
-  selectToEnd
-} = require('./shared.js');
+var selectionRange = require('selection-range');
 
 var ScribeTextBlockPlugin = function(block) {
   return function(scribe) {
@@ -22,9 +17,41 @@ var ScribeTextBlockPlugin = function(block) {
       }
     };
 
+    var selectToEnd = function() {
+      var selection = new scribe.api.Selection();
+      var range = selection.range.cloneRange();
+      range.setEndAfter(scribe.el.lastChild, 0);
+
+      return range;
+    };
+
+    var isAtStartOfBlock = function() {
+      var currentRange = selectionRange(scribe.el);
+
+      return (
+        currentRange.start === 0 &&
+        currentRange.end === 0 &&
+        currentRange.atStart
+      );
+    };
+
+    var getTotalLength = function() {
+      var selection = new scribe.api.Selection();
+      var range = selection.range.cloneRange();
+      range.selectNodeContents(scribe.el);
+
+      return range.toString().length;
+    };
+
+    var isAtEndOfBlock = function() {
+      var currentRange = selectionRange(scribe.el);
+
+      return (getTotalLength() === currentRange.end) && (currentRange.start === currentRange.end);
+    };
+
     var createBlocksFromParagraphs = function() {
       var fakeContent = document.createElement('div');
-      fakeContent.appendChild(selectToEnd(scribe).extractContents());
+      fakeContent.appendChild(selectToEnd().extractContents());
 
       stripFirstEmptyElement(fakeContent);
 
@@ -50,7 +77,7 @@ var ScribeTextBlockPlugin = function(block) {
       }
     };
 
-    var isAtStartBoolean = false;
+    var isAtStart = false;
 
     scribe.el.addEventListener('keydown', function(ev) {
 
@@ -61,10 +88,10 @@ var ScribeTextBlockPlugin = function(block) {
       if (ev.keyCode === 13 && !ev.shiftKey) { // enter pressed
         ev.preventDefault();
 
-        if (isAtEnd(scribe)) {
+        if (isAtEndOfBlock()) {
 
           // Remove any bad characters after current selection.
-          selectToEnd(scribe).extractContents();
+          selectToEnd().extractContents();
           block.mediator.trigger("block:create", 'Text', null, block.el, { autoFocus: true });
         } else {
           createBlocksFromParagraphs();
@@ -75,15 +102,15 @@ var ScribeTextBlockPlugin = function(block) {
           scribe.setContent('<p><br></p>');
         }
 
-      } else if ((ev.keyCode === 37 || ev.keyCode === 38) && isAtStart(scribe)) {
+      } else if ((ev.keyCode === 37 || ev.keyCode === 38) && isAtStartOfBlock()) {
         ev.preventDefault();
 
         block.mediator.trigger("block:focusPrevious", block.blockID);
-      } else if (ev.keyCode === 8 && isAtStart(scribe)) {
+      } else if (ev.keyCode === 8 && isAtStartOfBlock()) {
         ev.preventDefault();
 
-        isAtStartBoolean = true;
-      } else if ((ev.keyCode === 39 || ev.keyCode === 40) && isAtEnd(scribe)) {
+        isAtStart = true;
+      } else if ((ev.keyCode === 39 || ev.keyCode === 40) && isAtEndOfBlock()) {
         ev.preventDefault();
 
         block.mediator.trigger("block:focusNext", block.blockID);
@@ -96,14 +123,14 @@ var ScribeTextBlockPlugin = function(block) {
         return;
       }
 
-      if (ev.keyCode === 8 && isAtStartBoolean) {
+      if (ev.keyCode === 8 && isAtStart) {
         ev.preventDefault();
 
         block.mediator.trigger('block:remove', block.blockID, {
           transposeContent: true
         });
 
-        isAtStartBoolean = false;
+        isAtStart = false;
       }
     });
   };
